+++ 417-Pacific-Atlantic-Water-Flow/main.cpp
@@ -1,5 +1,6 @@
 #include <iostream>
 #include <vector>
+#include <queue>
 
 using namespace std;
 
@@ -14,103 +15,119 @@ public:
         int m = matrix.size();
         int n = matrix.back().size();
 
-        vector<vector<bool>> reachable_to_pacific(m, vector<bool>(n));
-        // from Pacific top-left to bottom-right
+        queue<pair<int, int>> cells_reachable;
+        vector<vector<bool>> reachable_to_atlantic(m, vector<bool>(n, false));
+
         for (int i = 0; i < m; ++i) {
-            for (int j = 0; j < n; ++j) {
-                if ((0 == i) || (0 == j)) {
-                    reachable_to_pacific[i][j] = true;
-                    continue;
-                }
-
-                reachable_to_pacific[i][j] = (reachable_to_pacific[i - 1][j] && (matrix[i][j] >= matrix[i - 1][j]));
-                reachable_to_pacific[i][j] = (reachable_to_pacific[i][j] ||
-                                              (reachable_to_pacific[i][j - 1] && (matrix[i][j] >= matrix[i][j - 1])));
-            }
+            cells_reachable.emplace(i, n - 1);
+            reachable_to_atlantic[i][n - 1] = true;
         }
-        // from Pacific bottom-right to top-left
-        for (int i = m - 1; i >= 0; --i) {
-            for (int j = n - 1; j >= 0; --j) {
-                if (reachable_to_pacific[i][j]) { continue; }
-
-                if ((j < n - 1) && reachable_to_pacific[i][j + 1] &&
-                    (matrix[i][j] >= matrix[i][j + 1])) {
-                    reachable_to_pacific[i][j] = true;
-                }
-
-                if (!reachable_to_pacific[i][j] && (i < m - 1) && reachable_to_pacific[i + 1][j] &&
-                    (matrix[i][j] >= matrix[i + 1][j])) {
-                    reachable_to_pacific[i][j] = true;
-                }
-            }
+        for (int i = 0; i < n - 1; ++i) {
+            cells_reachable.emplace(m - 1, i);
+            reachable_to_atlantic[m - 1][i] = true;
         }
 
-        /*
-        for (const auto &row : reachable_to_pacific) {
-            for (const auto &cell : row) {
-                cout << cell << " ";
-            }
-            cout << endl;
-        }
-        cout << "***" << endl;
-         */
+        // BFS to find out all cells can flow into the Pacific
+        while (!cells_reachable.empty()) {
+            const auto &pr = cells_reachable.front();
 
-        vector<vector<bool>> reachable_to_atlantic(m, vector<bool>(n));
-        // from Atlantic bottom-right to top-left
-        for (int i = m - 1; i >= 0; --i) {
-            for (int j = n - 1; j >= 0; --j) {
-                if ((m - 1 == i) || (n - 1 == j)) {
-                    reachable_to_atlantic[i][j] = true;
-                    continue;
-                }
+            int y, x;
+            // upper
+            y = pr.first - 1;
+            x = pr.second;
+            if ((y >= 0) && !reachable_to_atlantic[y][x] &&
+                (matrix[y][x] >= matrix[pr.first][pr.second])) {
+                cells_reachable.emplace(y, x);
+                reachable_to_atlantic[y][x] = true;
+            }
+            // right
+            y = pr.first;
+            x = pr.second + 1;
+            if ((x < n) && !reachable_to_atlantic[y][x] &&
+                (matrix[y][x] >= matrix[pr.first][pr.second])) {
+                cells_reachable.emplace(y, x);
+                reachable_to_atlantic[y][x] = true;
+            }
+            // bottom
+            y = pr.first + 1;
+            x = pr.second;
+            if ((y < m) && !reachable_to_atlantic[y][x] &&
+                (matrix[y][x] >= matrix[pr.first][pr.second])) {
+                cells_reachable.emplace(y, x);
+                reachable_to_atlantic[y][x] = true;
+            }
+            // left
+            y = pr.first;
+            x = pr.second - 1;
+            if ((x >= 0) && !reachable_to_atlantic[y][x] &&
+                (matrix[y][x] >= matrix[pr.first][pr.second])) {
+                cells_reachable.emplace(y, x);
+                reachable_to_atlantic[y][x] = true;
+            }
 
-                reachable_to_atlantic[i][j] = (reachable_to_atlantic[i + 1][j] && (matrix[i][j] >= matrix[i + 1][j]));
-                reachable_to_atlantic[i][j] = (reachable_to_atlantic[i][j] ||
-                                               (reachable_to_atlantic[i][j + 1] && (matrix[i][j] >= matrix[i][j + 1])));
-            }
+            cells_reachable.pop();
         }
-        /*
-        for (const auto &row : reachable_to_atlantic) {
-            for (const auto &cell : row) {
-                cout << cell << " ";
-            }
-            cout << endl;
-        }
-        cout << "***" << endl;
-         */
-        // from Atlantic top-left to bottom-right
+
+        // check for cells which can flow to the Pacific
+        vector<vector<bool>> reachable_to_pacific(m, vector<bool>(n, false));
         for (int i = 0; i < m; ++i) {
-            for (int j = 0; j < n; ++j) {
-                if (reachable_to_atlantic[i][j]) { continue; }
-
-                if ((j > 0) && reachable_to_atlantic[i][j - 1] &&
-                    (matrix[i][j] >= matrix[i][j - 1])) {
-                    reachable_to_atlantic[i][j] = true;
-                }
-
-                if (!reachable_to_atlantic[i][j] && (i > 0) && reachable_to_atlantic[i - 1][j] &&
-                    (matrix[i][j] >= matrix[i - 1][j])) {
-                    reachable_to_atlantic[i][j] = true;
-                }
-            }
+            cells_reachable.emplace(i, 0);
+            reachable_to_pacific[i][0] = true;
         }
-        /*
-        for (const auto &row : reachable_to_atlantic) {
-            for (const auto &cell : row) {
-                cout << cell << " ";
-            }
-            cout << endl;
+        for (int i = 1; i < n; ++i) {
+            cells_reachable.emplace(0, i);
+            reachable_to_pacific[0][i] = true;
         }
-        cout << "***" << endl;
-         */
 
-        // final round
-        for (int i = 0; i < m; ++i) {
-            for (int j = 0; j < n; ++j) {
-                if (reachable_to_pacific[i][j] && reachable_to_atlantic[i][j]) {
-                    grids.emplace_back(i, j);
-                }
+        while (!cells_reachable.empty()) {
+            const auto &pr = cells_reachable.front();
+
+            if (reachable_to_atlantic[pr.first][pr.second]) {
+                grids.emplace_back(pr.first, pr.second);
             }
+            //cout << "**" << endl;
+            //cout << "(" << pr.first << ", " << pr.second << ")" << endl;
+
+            int y, x;
+            // upper
+            y = pr.first - 1;
+            x = pr.second;
+            if ((y >= 0) && !reachable_to_pacific[y][x] &&
+                (matrix[y][x] >= matrix[pr.first][pr.second])) {
+                cells_reachable.emplace(y, x);
+
+                //cout << "\t(" << y << ", " << x << ")" << endl;
+                reachable_to_pacific[y][x] = true;
+            }
+            // right
+            y = pr.first;
+            x = pr.second + 1;
+            if ((x < n) && !reachable_to_pacific[y][x] &&
+                (matrix[y][x] >= matrix[pr.first][pr.second])) {
+                cells_reachable.emplace(y, x);
+                //cout << "\t(" << y << ", " << x << ")" << endl;
+                reachable_to_pacific[y][x] = true;
+            }
+            // bottom
+            y = pr.first + 1;
+            x = pr.second;
+            if ((y < m) && !reachable_to_pacific[y][x] &&
+                (matrix[y][x] >= matrix[pr.first][pr.second])) {
+                cells_reachable.emplace(y, x);
+                //cout << "\t(" << y << ", " << x << ")" << endl;
+                reachable_to_pacific[y][x] = true;
+            }
+            // left
+            y = pr.first;
+            x = pr.second - 1;
+            if ((x >= 0) && !reachable_to_pacific[y][x] &&
+                (matrix[y][x] >= matrix[pr.first][pr.second])) {
+                cells_reachable.emplace(y, x);
+                //cout << "\t(" << y << ", " << x << ")" << endl;
+                reachable_to_pacific[y][x] = true;
+            }
+
+            cells_reachable.pop();
         }
 
         return grids;
@@ -118,7 +135,6 @@ public:
 };
 
 int main() {
-    /*
     vector<vector<int>> matrix{
             {1, 2, 2, 3, 5},
             {3, 2, 3, 4, 4},
@@ -126,7 +142,6 @@ int main() {
             {6, 7, 1, 4, 5},
             {5, 1, 1, 2, 4},
     };
-     */
     /*
     vector<vector<int>> matrix{
             {1, 2, 3},
@@ -134,12 +149,14 @@ int main() {
             {7, 6, 5},
     };
      */
+    /*
     vector<vector<int>> matrix{
             {1,  2,  3,  4},
             {12, 13, 14, 5},
             {11, 16, 15, 6},
             {10, 9,  8,  7},
     };
+     */
 
     auto grids = Solution::pacificAtlantic(matrix);
     for (const auto &grid:grids) {